safari-534.53-branch/LayoutTests/canvas/philip/tests/toDataURL.arguments.2.html

<!DOCTYPE html>
<title>Canvas test: toDataURL.arguments.2</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL('image/png', 'another argument that should not raise an exception', 'and another');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");


});
</script>
