LayoutTests/canvas/philip/tests/toDataURL.arguments.2-expected.txt

< [index] >
toDataURL.arguments.2
toDataURL ignores extra arguments
References: toDataURL.arguments Bugzilla
Actual output:
Passed
